# Improve interpolate() speed for channels_last CPU videos